🎉 完成系统配置

+++ src/Memo.Blog.Application/Common/Models/PaginationResult.cs
@@ -7,10 +7,10 @@ public class PaginationResult<T>
     }
     public PaginationResult(IReadOnlyList<T> items, long total) : this(items)
     { 
-        Total = total;
+        Total = (int)total;
     }
 
-    public long Total { get; set; }
+    public int Total { get; set; }
     public IReadOnlyList<T> Items { get; set; }
     //public long Page { get; set; }
     //public long Size { get; set; }
